+++ data/crac/crac-api/src/test/java/com/farao_community/farao/data/crac_api/UnitTest.java
@@ -9,7 +9,6 @@ package com.farao_community.farao.data.crac_api;
 import com.farao_community.farao.commons.FaraoException;
 import org.junit.Test;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 
 /**
@@ -55,13 +54,5 @@ public class UnitTest {
         } catch (FaraoException e) {
             // should throw
         }
-    }
-
-    @Test
-    public void testLoopflowParameters() {
-        CnecLoopFlowExtension cnecLoopFlowExtension = new CnecLoopFlowExtension();
-        cnecLoopFlowExtension.setLoopFlowConstraint(100);
-        assertEquals(100, cnecLoopFlowExtension.getLoopFlowConstraint(), 0.1);
-        assertEquals(0.0, cnecLoopFlowExtension.getInputLoopFlow(), 0.1);
     }
 }
